Reorganized file hierarchy
