+++ code/part.txt
@@ -1,0 +1,3 @@
+2 15.0 0.01
+1.0 2.0 0.5 0.7 2
+4.0 8.0 0.1 -0.2 3
