+++ code/src/func.c
@@ -1,0 +1,7 @@
+#include "sim.h"
+
+
+int zoneDispute(node* native, node* insert)
+{
+  zone*
+}
