--- include/sim.h
@@ -1,10 +1,0 @@
-#include <stdlib.h>
-#include <stdio.h>
-#ifndef GALSIM
-/*
-  Barnes-Hut simulation algorithm
-  Relate forces by COM arguments, nlogn runtime
-
-*/
-
-#endif
