--- include/simConst.h
@@ -1,2 +1,0 @@
-extern float G = 6.67408e-11;//mks gravitational constant
-extern float pi = 3.14159274101257324;//maximum precision pi for 4B floating point
